refactor: TermType 클래스 분리
- `TermType` enum 클래스를 `LoginViewModel.kt`에서 `login/model/TermType.kt` 파일로 분리했습니다.
- `LoginScreen.kt`, `LoginViewModel.kt`, `PrivacyBottomSheet.kt`에서 분리된 `TermType` 클래스를 참조하도록 수정했습니다.
