+++ Near/app/src/main/java/com/alarmy/near/presentation/feature/login/LoginViewModel.kt
@@ -1,12 +1,10 @@
 package com.alarmy.near.presentation.feature.login
 
-import androidx.annotation.StringRes
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.alarmy.near.BuildConfig
-import com.alarmy.near.R
 import com.alarmy.near.data.repository.AuthRepository
 import com.alarmy.near.model.ProviderType
+import com.alarmy.near.presentation.feature.login.model.TermType
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -137,27 +135,6 @@ data class TermsAgreementState(
      */
     val isAllRequiredTermsAgreed: Boolean
         get() = isServiceTermsAgreed && isPrivacyCollectionAgreed && isPrivacyPolicyAgreed
-}
-
-/**
- * 약관 및 정책 타입
- */
-enum class TermType(
-    @StringRes val titleRes: Int,
-    val url: String,
-) {
-    SERVICE_TERMS(
-        titleRes = R.string.terms_service_agreed,
-        url = BuildConfig.SERVICE_AGREED_TERMS_URL,
-    ),
-    PRIVACY_COLLECTION(
-        titleRes = R.string.terms_personal_info,
-        url = BuildConfig.PERSONAL_INFO_TERMS_URL,
-    ),
-    PRIVACY_POLICY(
-        titleRes = R.string.terms_privacy_policy,
-        url = BuildConfig.PRIVACY_POLICY_TERMS_URL,
-    ),
 }
 
 /*
